Elimination System Prototype
To test the system: Scenes -> Testing -> Elimination System Prototype

It is currently wired to enemy death instead of player death as it was easier to test due to the reason that there is no enemies capable of doing damage.

An essential part of the system is the CreatureHealthData SO.
When creating a new enemy go to: DP source -> Data -> Enemies -> specific Enemy       and create a new SO: (right click in project folder) -> Data -> CreatureHealthData

InitMaxHealth specifies the base health of the creature.

Max Health Cap is so you can cap that specific creatures absolute maximum health cap.

Health Growth specifies the growth rate on player death.