I was working on fixing the multi-parallel reload glitch in the GunController script and this brought to my attention a more pressing and troublesome bug with the "accounting" done between the "Inventory.cs" and the "GunController.cs" script.
This led to some improvements in the "Awake() / Start()" of the Inventory,
as well as FIXING the very naive logic that was formerly found in the
"""public void DropItem(string name, int amount = 1, bool consume = false)""" function,
which was causing the "accounting" issue.

Now I will return to addressing the original Reload glitch, as intended.