Further 3rd person animation tweaks for shotgun timing.
Problem: The pistol and the rifle reload animation speed were pretty well in
sync, but the shotgun looked incredibly slow, compared to 1st person.

Solution: I set the shotgun to not "sync" with the "handgunReload" layer
so I could give it its own animation speed of 1.6, as opposed to 0.4,
used for handgun and rifle animation.

 Changes to be committed:
	modified:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Entities/Prisoner/PrisonerAimsetAnimationController.controller
	modified:   Devil's Punishment v2/Assets/Devil's Punishment Source/Scenes/Testing/Steamworks/GunDropFix.unity