Generator parts correctly positioned after inventory is sorted... caveat. Inventory is not being sorted after every exchange...
ex: Generally dropping an item will "scooch" all items after it "up" 1
index value. Leaving inventory misordered.
This needs to be addressed still in the next update.
Generally speaking the inventory is working functionally at this point
and successfully keeping track of the items.

One potential solution would be to "Sort()" after every exchange, but
this needs to be tested.

Another thing to consider is perhaps preventing more than 1
"GeneratorPart" of the same type from being picked up, from within the
"AddItem()" function of the Inventory, before the item makes its way
into the inventory's internal List.

I still haven't accounted for duplicate "GeneratorParts" of the same
type "on startup". I believe the outcome is that 1 of them would be
placed in the correct position, and the other would just be left alone
in it's same inventory spot (wherever it happens to be located).

 Changes to be committed:
	modified:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Entities/Prisoner/Player.prefab
	new file:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Items/Generator_Part_A.prefab
	new file:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Items/Generator_Part_A.prefab.meta
	new file:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Items/Generator_Part_B.prefab
	new file:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Items/Generator_Part_B.prefab.meta
	new file:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Items/Generator_Part_C.prefab
	new file:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Items/Generator_Part_C.prefab.meta
	modified:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/Items/ResourceManager.prefab
	modified:   Devil's Punishment v2/Assets/Devil's Punishment Source/Prefabs/UI/Inventory/GenSlot A.prefab
	modified:   Devil's Punishment v2/Assets/Devil's Punishment Source/Scenes/Testing/Steamworks/GunDropFix.unity
	modified:   Devil's Punishment v2/Assets/Devil's Punishment Source/Scripts/Items/Inventory.cs